add types.ts directory and refactor
